Move ScrollToEpisodesButton to new file
